flask_app/requirements.txt
streamlit==1.31.1
torch==2.5.1
torchvision==0.20.1
scikit-learn==1.3.1
scipy==1.11.3
pygame==2.5.2
Pillow==10.0.1
opencv-python-headless==4.8.1.78
flask
matplotlib
